refactor(modules): rename files to camelCase
affects: @action-land/component, @action-land/core, @action-land/tarz
