+++ modules/component/src/autoForward.ts
