+++ modules/component/src/component.ts
