+++ modules/core/index.ts
@@ -1,6 +1,6 @@
 export {action, IAction} from './src/action'
 export {isAction} from './src/isAction'
-export {Nil} from './src/Nil'
+export {Nil} from './src/nil'
 export {isNil} from './src/isNil'
-export {List} from './src/List'
+export {List} from './src/list'
 export {isList} from './src/isList'
