+++ modules/core/src/list.ts
@@ -1,6 +1,6 @@
 import {action, IAction} from './action'
 import {isNil} from './isNil'
-import {Nil} from './Nil'
+import {Nil} from './nil'
 
 /**
  * Action type for Nil
