+++ modules/tarz/index.ts
@@ -2,11 +2,11 @@
  * Created by tushar on 21/07/18
  */
 
-export {CommandFunction} from './src/CommandFunction'
+export {CommandFunction} from './src/commandFunction'
 export {concatC} from './src/concatC'
 export {concatR} from './src/concatR'
 export {matchC} from './src/matchC'
 export {matchR} from './src/matchR'
-export {ReducerFunction} from './src/ReducerFunction'
+export {ReducerFunction} from './src/reducerFunction'
 export {zeroC} from './src/zeroC'
 export {zeroR} from './src/zeroR'
