+++ modules/tarz/src/matchC.ts
@@ -5,7 +5,7 @@
 import {IAction, isAction, Nil} from '@action-land/core'
 import {CurriedFunction2, curry2} from 'ts-curry'
 
-import {CommandFunction} from './CommandFunction'
+import {CommandFunction} from './commandFunction'
 
 /**
  * Spec of Thunks for each Command that needs to be handled
