+++ modules/tarz/src/matchR.ts
@@ -5,7 +5,7 @@
 import {isAction} from '@action-land/core'
 import {CurriedFunction2, curry2} from 'ts-curry'
 
-import {ReducerFunction} from './ReducerFunction'
+import {ReducerFunction} from './reducerFunction'
 
 /**
  * Spec of Reducers for each Action that needs to be handled
