+++ test/component/autoForward.ts
