+++ test/component/component.ts
