+++ test/tarz/matchC.ts
@@ -5,7 +5,7 @@
 import {action, IAction, isNil} from '@action-land/core'
 import * as assert from 'assert'
 
-import {CommandFunction} from '../../modules/tarz/src/CommandFunction'
+import {CommandFunction} from '../../modules/tarz/src/commandFunction'
 import {matchC} from '../../modules/tarz/src/matchC'
 
 describe('matchC', () => {
